[ISSUE-45] Fix bug allowing library to automatically renew expired sessions
