+++ src/test/resources/mockResponses/errorInvalidApiKey.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<rsp stat="fail" version="1.0">
+    <err code="1">Invalid API key or user key</err>
+</rsp>
